server/python-bottle/requirements.txt
bottle==0.12.25
bottle-sqlalchemy==0.4.3
greenlet==2.0.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
